update env & raw model retrieval

+++ code/00_retrieve_data.py
@@ -45,18 +45,31 @@ def download_model_extracted():
 
 
 def download_model_raw():
-    cfg = nsl.io.load_config()
-    id = cfg.parameters.google_drive_ids.model_full.id
-    out_dir = cfg.model.input.full_model_run
-    out_dir.mkdir(parents=True, exist_ok=True)
-    tar_archive = out_dir.joinpath(cfg.parameters.google_drive_ids.model_full.file)
-    gdown.download(id=id, output=tar_archive.as_posix(), quiet=False)
-    # extract tar archive
-    tar = tarfile.open(tar_archive)
-    tar.extractall(path=out_dir)
-    tar.close()
-    # remove tar file
-    tar_archive.unlink()
+    """Raw model data are archived on the SIO server.
+    I had removed the data from google drive storage.
+    Please get in touch with Gunnar Voet gvoet@ucsd.edu if
+    you'd like to work with the raw model data.
+    """
+    if False:
+        cfg = nsl.io.load_config()
+        id = cfg.parameters.google_drive_ids.model_full.id
+        out_dir = cfg.model.input.full_model_run
+        out_dir.mkdir(parents=True, exist_ok=True)
+        tar_archive = out_dir.joinpath(cfg.parameters.google_drive_ids.model_full.file)
+        gdown.download(id=id, output=tar_archive.as_posix(), quiet=False)
+        # extract tar archive
+        tar = tarfile.open(tar_archive)
+        tar.extractall(path=out_dir)
+        tar.close()
+        # remove tar file
+        tar_archive.unlink()
+    else:
+        s = ("""Raw model data are archived on the SIO server.
+        Data were removed from google drive storage.
+        Please get in touch with Gunnar Voet gvoet@ucsd.edu if
+        you'd like to work with the raw model data.
+        """)
+        print(s)
 
 
 def clean_model_raw():
